+++ public/classes/models/listen/Event.js
@@ -10,9 +10,5 @@ export const ListerEventLi = {
   ...ListerLi,
   ArrowLeft: { nokey: 'leaveToParent', 'ctrl+maj': 'movePanelLeft' },
   m:         { meta:  'toggleDisplayMode' },
-  Tab:       { nokey: 'cycleLink', maj: 'cycleLinkBack' },
-  o:         { nokey: 'openActiveLink' },
-  g:         { nokey: 'goLink' },
-  a:         { nokey: 'splitLink' },
   t:         { nokey: 'openNaturePanel', meta: 'openToolsPanel' },
 }