+++ public/locale/fr/ERRORS.js
@@ -20,6 +20,7 @@ export const ERRORS = {
   
   /* -- Les CIBLES --*/
   , 5200: "Aucune cible n'est sélectionnée."
+  , 5210: "Cet élément ne possède aucune cible."
 
     /* -- 6000 : Double fenêtre  -- */
   , 6100: "La fenêtre n'est pas splitée. Pour la spliter : ⌥ + 2"